renames GoalEvent to PreconditionEvent deletes 2 unused subevents moves PreconditionEvents to their own package
