--- webofneeds/won-bot/src/main/java/won/bot/framework/eventbot/event/impl/analyzation/GoalShapeAmbivalentEvent.java
@@ -1,12 +1,0 @@
-package won.bot.framework.eventbot.event.impl.analyzation;
-
-import won.protocol.model.Connection;
-
-/**
- * Created by fsuda on 27.11.2017.
- */
-public class GoalShapeAmbivalentEvent extends GoalUnsatisfiedEvent {
-    public GoalShapeAmbivalentEvent(Connection con, Object payload) {
-        super(con, payload);
-    }
-}
