--- webofneeds/won-bot/src/main/java/won/bot/framework/eventbot/event/impl/analyzation/GoalShapeMissingEvent.java
@@ -1,12 +1,0 @@
-package won.bot.framework.eventbot.event.impl.analyzation;
-
-import won.protocol.model.Connection;
-
-/**
- * Created by fsuda on 27.11.2017.
- */
-public class GoalShapeMissingEvent extends GoalUnsatisfiedEvent{
-    public GoalShapeMissingEvent(Connection con, Object payload) {
-        super(con, payload);
-    }
-}
