+++ webofneeds/won-bot/src/main/java/won/bot/framework/eventbot/event/impl/analyzation/precondition/PreconditionEvent.java
@@ -1,4 +1,4 @@
-package won.bot.framework.eventbot.event.impl.analyzation;
+package won.bot.framework.eventbot.event.impl.analyzation.precondition;
 
 import won.bot.framework.eventbot.event.BaseNeedAndConnectionSpecificEvent;
 import won.protocol.model.Connection;
@@ -6,10 +6,10 @@ import won.protocol.model.Connection;
 /**
  * Created by fsuda on 27.11.2017.
  */
-public abstract class GoalEvent extends BaseNeedAndConnectionSpecificEvent {
+public abstract class PreconditionEvent extends BaseNeedAndConnectionSpecificEvent {
     private final Object payload;
 
-    public GoalEvent(Connection con, Object payload) {
+    public PreconditionEvent(Connection con, Object payload) {
         super(con);
         this.payload = payload;
     }
